Refactor y reorganiza módulos de utilidades y modelos externos para la API de SafeWaters
